SEQNG-371 Split SequenceToolbars into several source files
